moved peer list adapter to adapter folder
